Recebendo requisições do cliente

+++ servidor_pagamentos/src/main/java/com/sistemapag/service/IPagamentoService.java
@@ -5,5 +5,5 @@ import java.rmi.RemoteException;
 
 // processarPagamento retorna uma string indicando se o pagamento foi aprovado ou recusado.
 public interface IPagamentoService extends Remote { // Extende Remote para indicar que os métodos serão chamados remotamente
-    boolean processarPagamento(Long pedidoId, double valor) throws RemoteException;
+    boolean processarPagamento(int pedidoId, double valor) throws RemoteException;
 }
